+++ src/wdio/services/ServeStaticService.js
@@ -2,19 +2,95 @@ import serveStatic from '../../../scripts/serve/serve-static';
 import SERVICE_DEFAULTS from '../../../config/wdio/services.default-config';
 import Logger from '../../../scripts/utils/logger';
 
+const WebpackDevServer = require('webpack-dev-server');
+const webpack = require('webpack');
+
 const { serveStatic: SERVE_STATIC_DEFAULTS } = SERVICE_DEFAULTS;
 const context = '[Terra-Toolkit:serve-static-service]';
 
+const displayServer = () => {
+  Logger.log('Server started listening', { context });
+};
+
+// Create a new proxy watch function
+const watch = (compiler) => {
+  // Store off original watch function.
+  const origWatch = compiler.watch;
+  // Return new watch function
+  return (watchOptions, handler) => {
+    // Call origial watch functions with the compiler as 'this'
+    const watcher = origWatch.call(compiler, watchOptions, handler);
+    // Remove the 'watch' function from the returned watcher.
+    watcher.watch = () => {
+      // console.log('not watching');
+      Logger.log('Watching is disabled', { context });
+    };
+    return watcher;
+  };
+};
+
+// Create a webpack dev server instance.
+const startWebpackDevServer = (options) => {
+  const {
+    port, index, locale,
+  } = options;
+  const host = '0.0.0.0';
+  let { config } = options;
+  // if config is a function, execute it with prod mode if applicable.
+  if (typeof config === 'function') {
+    const env = { ...locale && { defaultLocale: locale } };
+    config = config(env, { p: true });
+  }
+
+  // pull the dev server options out of the webpack config. override host, port, and stats. SRY.
+  const devServerOptions = Object.assign({}, config.devServer, {
+    // Disable hot reloading
+    hot: false,
+    inline: false,
+    host,
+    port,
+    index,
+    stats: {
+      colors: true,
+      children: false,
+    },
+  });
+
+  return new Promise((resolve, reject) => {
+    // get a compiler
+    const compiler = webpack(config);
+    // Disable hot reloading
+    compiler.watch = watch(compiler);
+    // get a server
+    const devServer = new WebpackDevServer(compiler, devServerOptions);
+    // add a hook to report when webpacking is done
+    compiler.hooks.done.tap('Done', (stats) => {
+      if (stats.hasErrors()) {
+        reject();
+      } else {
+        resolve(devServer);
+      }
+    });
+
+    // start that server.
+    devServer.listen(port, host, (err) => {
+      if (err) {
+        reject(err);
+      }
+      displayServer();
+    });
+  });
+};
+
 export default class ServeStaticService {
   async onPrepare(config = {}) {
-    const { site, webpackConfig, locale } = config;
+    const { site, webpackConfig, locale = process.env.LOCALE } = config;
 
     if (!webpackConfig && !site) {
       Logger.warn('No webpack configuration provided', { context });
       return;
     }
 
-    const verbose = config.logLevel !== 'silent';
     const port = (config.serveStatic || {}).port || SERVE_STATIC_DEFAULTS.port;
     const index = (config.serveStatic || {}).index || SERVE_STATIC_DEFAULTS.index;
     // Explicitly not providing a fallback locale. Providing a fallback will lock the locale for all test runs when using the tt-wdio-runner.
@@ -25,12 +101,12 @@ export default class ServeStaticService {
     }
 
     const serveOptions = {
-      ...site && { site, disk: true },
+      ...site && { site },
       config: webpackConfig,
       port,
       index,
       locale,
-      verbose,
+      production: true,
     };
 
     await ServeStaticService.startService(serveOptions).then((server) => {
@@ -42,11 +118,12 @@ export default class ServeStaticService {
     await this.stop();
   }
 
-  // Options include config, site, port, index, locale, verbose
+  // Options include config, site, port, index, locale
   static startService(serveOptions) {
-    return serveStatic({
-      ...serveOptions, production: true,
-    });
+    if (serveOptions.site) {
+      return serveStatic(serveOptions);
+    }
+    return startWebpackDevServer(serveOptions);
   }
 
   stop() {